peridot additions:
everything should work as before with no breakage

can now use a monster[int] map of monsters for selection priority

if no monster in the map of monsters is found in the choice adventure, will list the monsters that were available as part of the error

can omit location if already in the choice adventure (e.g. in a choice adventure script)

can feed the choice adventure page to the function if already have it from elsewhere to save a server hit (e.g. in a choice adventure script)